Fix ICC warning about change of sign
debuggerstartparameters.h(86): warning #68: integer conversion resulted in a change of sign

-1 cannot be stored in a quint16. So change the port variable from
 quint16 to int.

+++ src/plugins/debugger/debuggerstartparameters.h
@@ -74,7 +74,7 @@ public:
 
     // Used by Qml debugging.
     QString qmlServerAddress;
-    quint16 qmlServerPort = InvalidPort;
+    int qmlServerPort = InvalidPort;
 
     // Used by general remote debugging.
     QString remoteChannel;
